refactored to not use getserversideprops
